relief_campaign_service/requirements.txt
Django==4.1
psycopg2-binary==2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.13.0
requests
celery
redis
pika